Try to reset the timestamp on the release
Prior to this commit, the releases are getting updated as expected.
However, it continues to give a "released on" date from the very initial
release creation. Sending a release into a draft and back to published
will reset this date and we can set it to a draft by deleting the tag
out from under it. What is unclear until GH actions run this commit, is
whether the following steps will re-publish it or create a new one,
leaving this now-draft as a redundant leftover.