Suppress an unhelpful clippy warning
The .filter() on an Iterator will always double-ref in this case. There
is no more efficient way to avoid the double-ref here. The best we could
do here is manually (or with metadata) inline these functions.

I want to keep these functions extracted for maintainability reasons and
I trust the compiler to inline these in this case. If that doesn't turn
out to be the case, this isn't particularly performance critical.