Fix issue with unreliable indexer updates
Fixes #1

We wrap the indexer's ID in `{` and `}` when saving to Sonarr. When
searching for indexer candidates to _update_ we were not searching with
those delimiters, which meant the syncer was susceptible to getting a
false positive match.

When there were two indexers where one's identifier was a substring of
another, an indexer update could fail depending on the order they were
processed.

For instance, in Jackett there are the following two indexer IDs:

  - `kickasstorrents`
  - `kickasstorrents-kathow`

If Sonarr returned the second one first in its API for existing
indexers, errors would occur.

When `kickasstorrents` was processed first by the syncer, it would try
to update `kickasstorrents-kathow`, by renaming it to the full name for
`kickasstorrents`. Sonarr rejected this for not having a unique name.

Furthermore, while `kickasstorrents-kathow` was not overwritten in
Sonarr, it would be overwritten locally, which could then cause later
mismatches when finding the indexer to update later on during the sync.